+++ src/static/js/Changeset.js
@@ -1806,16 +1806,15 @@ exports.copyAText = (atext1, atext2) => {
 };
 
 /**
- * Append the set of operations from atext to an assembler.
- *
- * @param {AText} atext -
- * @param assem - Assembler like SmartOpAssembler TODO add desc
+ * Convert AText to a series of operations.
+ * @param {AText} atext - The AText to convert.
+ * @yields {Op}
  */
-exports.appendATextToAssembler = (atext, assem) => {
+exports.opsFromAText = function* (atext) {
   // intentionally skips last newline char of atext
   let lastOp = null;
   for (const op of new exports.OpIter(atext.attribs)) {
-    if (lastOp != null) assem.append(lastOp);
+    if (lastOp != null) yield lastOp;
     lastOp = op;
   }
   if (lastOp == null) return;
@@ -1828,11 +1827,24 @@ exports.appendATextToAssembler = (atext, assem) => {
     const lastLineLength = atext.text.length - nextToLastNewlineEnd - 1;
     lastOp.lines--;
     lastOp.chars -= (lastLineLength + 1);
-    assem.append(lastOp);
+    yield copyOp(lastOp);
     lastOp.lines = 0;
     lastOp.chars = lastLineLength;
   }
-  if (lastOp.chars) assem.append(lastOp);
+  if (lastOp.chars) yield lastOp;
+};
+
+/**
+ * Append the set of operations from atext to an assembler.
+ *
+ * @deprecated Use `opsFromAText` instead.
+ * @param {AText} atext -
+ * @param assem - Assembler like SmartOpAssembler TODO add desc
+ */
+exports.appendATextToAssembler = (atext, assem) => {
+  warnDeprecated(
+      'Changeset.appendATextToAssembler() is deprecated; use Changeset.opsFromAText() instead');
+  for (const op of exports.opsFromAText(atext)) assem.append(op);
 };
 
 /**